unitTest

+++ src/main/java/net/metrosystems/mylibrary23/data/repository/LibraryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LibraryRepository extends CrudRepository<Library, Integer> {
+public interface LibraryRepository extends CrudRepository<Library, Long> {
     Library findByNameAndAddress(String name, String address);
 
 }
